+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/InstrumentService.java
@@ -3,12 +3,18 @@ package uk.ac.bbsrc.tgac.miso.core.service;
 import java.io.IOException;
 import java.util.List;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Barcodable.EntityType;
 import uk.ac.bbsrc.tgac.miso.core.data.Instrument;
 import uk.ac.bbsrc.tgac.miso.core.data.type.InstrumentType;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
-public interface InstrumentService extends DeleterService<Instrument>, ListService<Instrument>, PaginatedDataSource<Instrument>,
-    SaveService<Instrument> {
+public interface InstrumentService extends BarcodableService<Instrument>, DeleterService<Instrument>, ListService<Instrument>,
+    PaginatedDataSource<Instrument>, SaveService<Instrument> {
+
+  @Override
+  public default EntityType getEntityType() {
+    return EntityType.INSTRUMENT;
+  }
 
   List<Instrument> listByType(InstrumentType type) throws IOException;