+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/WorkstationService.java
@@ -1,7 +1,14 @@
 package uk.ac.bbsrc.tgac.miso.core.service;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Barcodable.EntityType;
 import uk.ac.bbsrc.tgac.miso.core.data.Workstation;
 
-public interface WorkstationService extends DeleterService<Workstation>, ListService<Workstation>, SaveService<Workstation> {
+public interface WorkstationService
+    extends BarcodableService<Workstation>, DeleterService<Workstation>, ListService<Workstation>, BulkSaveService<Workstation> {
+
+  @Override
+  public default EntityType getEntityType() {
+    return EntityType.WORKSTATION;
+  }
 
 }