+++ miso-web/src/main/webapp/scripts/lims.js
@@ -1094,7 +1094,7 @@ Utils.validation = {
   dateRegex: '^(19|20)[0-9]{2}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])$',
   dateTimeRegex: '^(19|20)[0-9]{2}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01]) ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$',
   sanitizeRegex: '^[^<>&\\t\\n\\r]*$',
-  uriComponentRegex: '[^<>&%;\/\\\\]*$', // problem characters even when encoded in URI components: %;\/
+  uriComponentRegex: '^[^<>&%;\/\\\\]*$', // problem characters even when encoded in URI components: %;\/
   alphanumRegex: '^[-_\\w]*$',
   unicodeWordRegex: '^[\\p{L}0-9_\\^\\-\\.\\s]+$',
   _unicodeWord: XRegExp('^[\\p{L}0-9_\\^\\-\\.\\s]+$'),