+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/WorkstationDao.java
@@ -1,13 +1,17 @@
 package uk.ac.bbsrc.tgac.miso.persistence;
 
 import java.io.IOException;
+import java.util.Collection;
+import java.util.List;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Workstation;
 
 public interface WorkstationDao extends SaveDao<Workstation> {
 
-  public Workstation getByAlias(String alias) throws IOException;
+  Workstation getByAlias(String alias) throws IOException;
 
-  public long getUsage(Workstation workstation) throws IOException;
+  long getUsage(Workstation workstation) throws IOException;
+
+  List<Workstation> listByIdList(Collection<Long> ids) throws IOException;
 
 }